# Fix bug in get enricher