This first modifies the benchmark `BenchmarkMetaTagEnricher` so it also tests obtaining the enricher, then it changes the getEnricher() method so that it can actually reuse the instantiated enricher.
```
replay@mst-nb:~/go/src/github.com/grafana/metrictank/idx/memory$ benchcmp /tmp/bench1 /tmp/bench2
benchmark                       old ns/op     new ns/op     delta
BenchmarkMetaTagEnricher-12     2006728       792810        -60.49%

benchmark                       old allocs     new allocs     delta
BenchmarkMetaTagEnricher-12     27022          2045           -92.43%

benchmark                       old bytes     new bytes     delta
BenchmarkMetaTagEnricher-12     985807        103507        -89.50%
```